@@ -12,51 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.5",
+    "@angular/animations": "^5.0.0-rc.3",
     "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.4.5",
-    "@angular/compiler": "^4.4.5",
-    "@angular/core": "^4.4.5",
-    "@angular/forms": "^4.4.5",
-    "@angular/http": "^4.4.5",
+    "@angular/common": "^5.0.0-rc.3",
+    "@angular/compiler": "^5.0.0-rc.3",
+    "@angular/core": "^5.0.0-rc.3",
+    "@angular/forms": "^5.0.0-rc.3",
+    "@angular/http": "^5.0.0-rc.3",
     "@angular/material": "2.0.0-beta.12",
-    "@angular/platform-browser": "^4.4.5",
-    "@angular/platform-browser-dynamic": "^4.4.5",
-    "@angular/router": "^4.4.5",
-    "@ngrx/db": "^2.0.1",
-    "@ngrx/effects": "^4.0.5",
-    "@ngrx/router-store": "^4.0.4",
-    "@ngrx/store": "^4.0.3",
-    "@ngrx/store-devtools": "^4.0.0",
-    "angularfire2": "^5.0.0-rc.3",
+    "@angular/platform-browser": "^5.0.0-rc.3",
+    "@angular/platform-browser-dynamic": "^5.0.0-rc.3",
+    "@angular/router": "^5.0.0-rc.3",
+    "@ngrx/effects": "^4.1.0",
+    "@ngrx/router-store": "^4.1.0",
+    "@ngrx/store": "^4.1.0",
+    "@ngrx/store-devtools": "github:ngrx/store-devtools-builds",
+    "angularfire2": "5.0.0-rc.3",
     "core-js": "^2.5.1",
-    "firebase": "^4.5.2",
-    "firebaseui": "^2.4.0",
-    "firebaseui-angular": "^0.4.5",
+    "firebase": "^4.6.0",
+    "firebaseui": "^2.4.1",
+    "firebaseui-angular": "^2.4.0",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.4.2",
+    "rxjs": "^5.5.0",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.7",
-    "@angular/compiler-cli": "^4.4.5",
-    "@angular/language-service": "^4.4.5",
+    "@angular/cli": "1.4.9",
+    "@angular/compiler-cli": "^5.0.0-rc.3",
+    "@angular/language-service": "^4.4.6",
     "@types/jasmine": "~2.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.0.34",
+    "@types/node": "~8.0.46",
     "codelyzer": "~3.2.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "~5.2.0",
     "ts-node": "~3.3.0",
-    "tslint": "~5.7.0",
+    "tslint": "~5.8.0",
     "typescript": "~2.5.3"
   }
 }
